+++ src/main/java/com/hearthsim/card/spellcard/concrete/TheCoin.java
@@ -51,8 +51,8 @@ public class TheCoin extends SpellCard {
 	{
 		HearthTreeNode toRet = super.use_core(side, targetMinion, boardState, deckPlayer0, deckPlayer1, singleRealizationOnly);
 		if (toRet != null) {
-			int newMana = toRet.data_.getCurrentPlayer().getMana();
-			newMana = newMana >= 10 ? newMana : newMana + 1;
+			byte newMana = toRet.data_.getCurrentPlayer().getMana();
+			newMana = newMana >= 10 ? newMana : (byte)(newMana + 1);
 			toRet.data_.getCurrentPlayer().setMana(newMana);
 		}
 		return boardState;
